Hide default multiplexer impl into cpp file
