+++ libcaf_net/caf/net/fwd.hpp
@@ -90,12 +90,6 @@ enum class errc;
 
 } // namespace caf::net::octet_stream
 
-namespace caf::detail {
-
-class pollset_updater;
-
-} // namespace caf::detail
-
 namespace caf::net::lp {
 
 class client;
